changed seedRandomGenerator parameter type from float to int

+++ headers/SCETools.hpp
@@ -31,7 +31,7 @@ namespace Math
 {
     float lerp(float a, float b, float v);
     float mapToRange(float fromMin, float fromMax, float toMin, float toMax, float val);
-    void seedRandomGenerator(float val);
+    void seedRandomGenerator(int val);
     float randRange(float min, float max);
 }
